refactor jwt verifiable credential tests
